Move DOI page to templates
