+++ templates/doi/cards/index.js
@@ -1,0 +1,3 @@
+export CoverageCard from './coverage-card'
+export ExportCard from './export-card'
+export TableCard from './table-card'
